requirements.txt
backcall==0.2.0
beautifulsoup4==4.9.1
boto==2.49.0
boto3==1.14.63
botocore==1.17.63
bpemb==0.3.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1
decorator==4.4.2
docutils==0.15.2
elasticsearch==7.9.1
future==0.18.2
gensim==3.8.3
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
joblib==0.16.0
nltk==3.5
nmslib==2.0.6
numpy==1.19.2
pandas==1.1.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
psutil==5.7.2
ptyprocess==0.6.0
pybind11==2.5.0
pycparser==2.20
Pygments==2.7.0
python-dateutil==2.8.1
python-telegram-bot==12.8
pytorch-pretrained-bert==0.6.2
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
sentencepiece==0.1.92
six==1.15.0
sklearn==0.0
smart-open==2.1.1
soupsieve==2.0.1
threadpoolctl==2.1.0
torch==1.6.0
tornado==6.0.4
tqdm==4.49.0
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
deelpavlov==0.12.1

